backend/requirements.txt
alembic==1.5.5
appdirs==1.4.4
astor==0.8.1
astroid==2.5.1
autopep8==1.5.5
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==5.0.7
et-xmlfile==1.0.1
evdev==1.4.0
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
lxml==4.6.2
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.3
pandasgui==0.2.10.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyarrow==4.0.0
pycodestyle==2.6.0
Pygments==2.8.1
pylint==2.7.2
pynput==1.7.3
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-xlib==0.29
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.23
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
wordcloud==1.8.1
wrapt==1.12.1
xlrd==2.0.1
